Change project structure
Now src is not compiled as a lib. Instead, the whole project is compiled
as a single executable. The tests project is also a single project now
with its own main file, instead of different separated project for each
test class.

One flaw of this new approach is that files in src must be included
manually again in the tests projects, so this still needs some improvement.